# src/com.wxw/04_project_actual/06_actual_swagger/api/user.go
/*
@Time : 2022/5/13 15:34
@Author : weixiaowei
@File : user
*/
// Package api defines the user model.
package api

// User represents body of User request and response.
type User struct {
	// User's name.
	// Required: true
	Name string `json:"name"`

	// User's nickname.
	// Required: true
	Nickname string `json:"nickname"`

	// User's address.
	Address string `json:"address"`

	// User's email.
	Email string `json:"email"`
}